fix(dev):仓库配送量和自提量修改

+++ o2-metadata-management/src/main/java/org/o2/metadata/console/infra/redis/WarehouseRedis.java
@@ -35,7 +35,7 @@ public interface WarehouseRedis {
      * @param expressQuantity 快递配送接单量限制
      * @return -1 失败 1成功
      */
-    Integer updateExpressQuantity(String warehouseCode, String expressQuantity, Long tenantId);
+    Long updateExpressQuantity(String warehouseCode, String expressQuantity, Long tenantId);
 
     /**
      * 仓库自提接单量增量更新
@@ -45,7 +45,7 @@ public interface WarehouseRedis {
      * @param pickUpQuantity     自提单量增量
      * @return 1 成功 -1 失败
      */
-    Integer updatePickUpValue(String warehouseCode, String pickUpQuantity, Long tenantId);
+    Long updatePickUpValue(String warehouseCode, String pickUpQuantity, Long tenantId);
